research/ml/requirements.txt
openai>=1.30.0
transformers>=4.40.0
peft>=0.10.0
datasets>=2.18.0
pillow>=10.3.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
torch>=2.2.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
huggingface-hub>=0.22.0
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0
onnxruntime>=1.17.0
ultralytics>=8.1.0
opencv-python>=4.9.0
albumentations>=1.3.0
timm>=0.9.0
coremltools>=7.0
scikit-image>=0.22.0
seaborn>=0.13.0
scipy>=1.12.0
gdown>=5.1.0
